+++ src/Resources/TrackTrace.php
@@ -12,6 +12,12 @@ class TrackTrace extends Trace
     /** @var bool */
     public $isDelivered;
 
+    /** @var string */
+    public $link;
+
+    /** @var string */
+    public $link_portal;
+
     /** @var array */
     public $history;
 
@@ -42,6 +48,25 @@ class TrackTrace extends Trace
             });
     }
 
+    public function getLinkConsumerPortalAttribute(): string
+    {
+        return $this->link_portal;
+    }
+
+    public function setLinkConsumerPortalAttribute(string $value): void
+    {
+        $this->link_portal = $value;
+    }
+
+    public function getLinkTracktraceAttribute(): string
+    {
+        return $this->link;
+    }
+
+    public function setLinkTracktraceAttribute(string $value): void
+    {
+        $this->link = $value;
+    }
     public function setHistoryAttribute(array $array = []): void
     {
         $this->history = collect($array)->map(function ($item) {